requirements.txt
glfw==2.8.0
moderngl==5.12.0
pyglm==2.8.0
pillow==11.1.0
opensimplex==0.4.5.1
